refactor supply_costs.py
